Modification de lfork poru coller a fork

+++ vm/src/cw_ins_lfork.c
@@ -6,7 +6,7 @@
 /*   By: ademenet <ademenet@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2016/07/21 16:07:20 by ademenet          #+#    #+#             */
-/*   Updated: 2016/08/08 18:00:01 by ademenet         ###   ########.fr       */
+/*   Updated: 2016/08/12 16:13:33 by ademenet         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -20,11 +20,11 @@ int			cw_ins_lfork(t_proc *proc, t_champion *tmp, t_ocp *ocp)
 {
 	unsigned int	p;
 
-	p = cw_get_data_dir(proc, tmp, (tmp->pc + 1) % MEM_SIZE, 2);
+	p = (short)cw_get_data_dir(proc, tmp, (tmp->pc + 1) % MEM_SIZE, 2);
 	cw_lst_push(&proc->champions, cw_lst_new(tmp->header, tmp->num));
 	proc->champions->ins = NULL;
 	proc->champions->pc_origin = tmp->pc_origin;
-	proc->champions->pc = (tmp->pc + p) % MEM_SIZE;
+	proc->champions->pc = (tmp->pc + (unsigned short)p) % MEM_SIZE;
 	proc->champions->carry = tmp->carry;
 	proc->champions->inst_c = 0;
 	proc->champions->lives = 0;
